Index and app separated, utils added for config
